requirements.txt
nonebot2>=2.0.0
nonebot-adapter-onebot>=2.2.0
aiohttp>=3.8.1
python-dateutil>=2.8.2
pydantic>=1.10.0
typing-extensions>=4.0.0
openai>=1.3.0
httpx>=0.24.1
jmespath>=1.0.1
python-docx>=0.8.11
openpyxl>=3.0.9
Pillow>=9.0.0 